issue-634: refactor(meal-planner): extract meal_planner trailing widgets (#634)
Reduce meal_planner_screen.dart 2968 -> 2055 lines by moving the trailing
widget classes into lib/widgets/meal_planner/ (verbatim; public where the
screen references them, Dart privacy is per-file):

- meal_day_card.dart            (DayCard; imports meal_badges)
- meal_badges.dart              (Stars + NutriBadge, used only by DayCard)
- meal_swap_sheet.dart          (SwapSheet)
- meal_consolidated_sheet.dart  (ConsolidatedSheet)
- meal_weekly_summary_card.dart (WeeklySummaryCard)

All extracted files <500 lines. Removed now-unused screen imports
(star_rating_row, pantry_coverage_badge). Residual >500 is the
_MealPlannerScreenState body; clean reduction needs #632 (per #634 note).
Verified: analyze clean on changed files (baseline 81 unchanged); 219 tests
pass (screens + meal_planner_generation + meal_generation_validation).

Part of #634 (meal_planner; dashboard remains — held for #1031 conflict).
