+++ monthy_budget_flutter/lib/screens/meal_planner_screen.dart
@@ -20,16 +20,18 @@ import '../widgets/freeform_meal_sheet.dart';
 import '../models/meal_budget_insight.dart';
 import '../utils/meal_budget_insights.dart';
 import '../widgets/meal_cost_reconciliation_sheet.dart';
-import '../widgets/star_rating_row.dart';
 import '../widgets/meal_plan_budget_sheet.dart';
 import '../widgets/nutrition_dashboard_card.dart';
 import '../onboarding/meals_tour.dart';
 import '../utils/pantry_matching.dart';
 import '../utils/rate_limiter.dart';
 import '../utils/waste_calculator.dart';
-import '../widgets/pantry_coverage_badge.dart';
 import '../widgets/pantry_quick_picker_sheet.dart';
 import 'meal_wizard_screen.dart';
+import '../widgets/meal_planner/meal_day_card.dart';
+import '../widgets/meal_planner/meal_swap_sheet.dart';
+import '../widgets/meal_planner/meal_consolidated_sheet.dart';
+import '../widgets/meal_planner/meal_weekly_summary_card.dart';
 
 class MealPlannerScreen extends StatefulWidget {
   final AppSettings settings;
@@ -1150,7 +1152,7 @@ class _MealPlannerScreenState extends State<MealPlannerScreen> {
 
     CalmBottomSheet.show(
       context,
-      builder: (_) => _SwapSheet(
+      builder: (_) => SwapSheet(
         currentRecipeId: currentRecipeId,
         nPessoas: plan.nPessoas,
         ingredientMap: iMap,
@@ -1181,7 +1183,7 @@ class _MealPlannerScreenState extends State<MealPlannerScreen> {
 
     CalmBottomSheet.show(
       context,
-      builder: (_) => _SwapSheet(
+      builder: (_) => SwapSheet(
         currentRecipeId: currentRecipeId,
         nPessoas: plan.nPessoas,
         ingredientMap: iMap,
@@ -1284,7 +1286,7 @@ class _MealPlannerScreenState extends State<MealPlannerScreen> {
 
     CalmBottomSheet.show(
       context,
-      builder: (_) => _ConsolidatedSheet(
+      builder: (_) => ConsolidatedSheet(
         totals: totals,
         ingredientMap: iMap,
         onAddToShoppingList: widget.onAddToShoppingList,
@@ -1829,7 +1831,7 @@ class _MealPlannerScreenState extends State<MealPlannerScreen> {
                                 settings: widget.settings.mealSettings,
                               ),
                               if (_weeklySummaries.containsKey(_selectedWeek))
-                                _WeeklySummaryCard(
+                                WeeklySummaryCard(
                                   summary: _weeklySummaries[_selectedWeek]!,
                                 ),
                             ],
@@ -1952,7 +1954,7 @@ class _MealPlannerScreenState extends State<MealPlannerScreen> {
                           ),
                         );
                       }
-                      return _DayCard(
+                      return DayCard(
                         mealDay: mainDay,
                         allCourses: courses,
                         plan: plan,
@@ -2049,920 +2051,5 @@ class _MealPlannerScreenState extends State<MealPlannerScreen> {
     return plan.days
         .where((d) => d.dayIndex >= start && d.dayIndex <= end)
         .toList();
-  }
-}
-
-// -- Day Card -----------------------------------------------------------------
-
-class _DayCard extends StatelessWidget {
-  final MealDay mealDay;
-
-  /// All courses for this day+mealType group (soup, main, dessert).
-  final List<MealDay> allCourses;
-  final MealPlan plan;
-  final MealPlannerService service;
-  final RecipeAiContent? aiContent;
-  final bool isExpanded;
-  final VoidCallback onToggleExpand;
-  final VoidCallback onSwap;
-  final void Function(MealDay day) onSwapCourse;
-  final VoidCallback onReplaceFreeform;
-  final void Function(ShoppingItem) onAddIngredientToList;
-  final ValueChanged<int> onRating;
-  final VoidCallback onViewPrepGuide;
-  final Set<String> activePantryIds;
-  final void Function(String ingredientId) onSubstituteIngredient;
-  final void Function(MealDay day, String ingredientId)
-  onSubstituteCourseIngredient;
-
-  const _DayCard({
-    required this.mealDay,
-    this.allCourses = const [],
-    required this.plan,
-    required this.service,
-    required this.aiContent,
-    required this.isExpanded,
-    required this.onToggleExpand,
-    required this.onSwap,
-    required this.onSwapCourse,
-    required this.onReplaceFreeform,
-    required this.onAddIngredientToList,
-    required this.onRating,
-    required this.onViewPrepGuide,
-    required this.onSubstituteIngredient,
-    required this.onSubstituteCourseIngredient,
-    this.activePantryIds = const {},
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = S.of(context);
-    final recipe = service.recipeMap[mealDay.recipeId];
-    if (recipe == null) return const SizedBox();
-    final iMap = service.ingredientMap;
-    // Total cost for all courses in this meal slot
-    final totalMealCost = allCourses.fold(
-      0.0,
-      (sum, c) => sum + c.costEstimate,
-    );
-    final displayCost = allCourses.length > 1
-        ? totalMealCost
-        : mealDay.costEstimate;
-    final costPerPerson = plan.nPessoas > 0
-        ? displayCost / plan.nPessoas
-        : displayCost;
-
-    return CalmCard(
-      padding: const EdgeInsets.fromLTRB(16, 12, 16, 12),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Day header: eyebrow + meal-type pill + leftover pill + cost + menu
-          CalmEyebrow(l10n.mealDayLabel(mealDay.dayIndex)),
-          const SizedBox(height: 6),
-          Row(
-            children: [
-              CalmPill(
-                label: mealDay.mealType.localizedLabel(l10n),
-                color: AppColors.ok(context),
-              ),
-              if (mealDay.isLeftover) ...[
-                const SizedBox(width: 4),
-                CalmPill(
-                  label: l10n.mealLeftover,
-                  color: AppColors.warn(context),
-                ),
-              ],
-              if (activePantryIds.isNotEmpty) ...[
-                const SizedBox(width: 4),
-                PantryCoverageBadge(
-                  coverageRatio: computePantryCoverage(
-                    recipe,
-                    activePantryIds,
-                  ).coverageRatio,
-                ),
-              ],
-              const Spacer(),
-              Text(
-                '${displayCost.toStringAsFixed(2)}${currencySymbol()}',
-                style: TextStyle(
-                  fontSize: 13,
-                  fontWeight: FontWeight.w700,
-                  color: AppColors.textPrimary(context),
-                ),
-              ),
-              PopupMenuButton<String>(
-                icon: Icon(
-                  Icons.more_vert,
-                  size: 18,
-                  color: AppColors.textMuted(context),
-                ),
-                padding: EdgeInsets.zero,
-                itemBuilder: (_) => [
-                  PopupMenuItem(
-                    value: 'freeform',
-                    child: Text(l10n.freeformReplace),
-                  ),
-                ],
-                onSelected: (value) {
-                  if (value == 'freeform') onReplaceFreeform();
-                },
-              ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          // Course rows as CalmListTile
-          ..._buildCourseListTiles(context, l10n, recipe, iMap),
-          const SizedBox(height: 4),
-          // Compact info row: stars + time + cost/person
-          Row(
-            children: [
-              _Stars(recipe.complexity),
-              const SizedBox(width: 8),
-              Icon(
-                Icons.timer_outlined,
-                size: 13,
-                color: AppColors.textMuted(context),
-              ),
-              const SizedBox(width: 2),
-              Text(
-                recipe.activeMinutes != null && recipe.passiveMinutes != null
-                    ? '${recipe.activeMinutes}+${recipe.passiveMinutes}min'
-                    : '${recipe.prepMinutes}min',
-                style: TextStyle(
-                  fontSize: 11,
-                  color: AppColors.textMuted(context),
-                ),
-              ),
-              const SizedBox(width: 8),
-              Icon(
-                Icons.person_outline,
-                size: 13,
-                color: AppColors.textMuted(context),
-              ),
-              const SizedBox(width: 2),
-              Text(
-                l10n.mealCostPerPerson(costPerPerson.toStringAsFixed(2)),
-                style: TextStyle(
-                  fontSize: 11,
-                  color: AppColors.textMuted(context),
-                ),
-              ),
-            ],
-          ),
-          // Compact nutrition badges
-          if (recipe.nutrition != null) ...[
-            const SizedBox(height: 4),
-            Row(
-              children: [
-                _NutriBadge(
-                  '${recipe.nutrition!.kcal}',
-                  'kcal',
-                  AppColors.error(context),
-                ),
-                const SizedBox(width: 4),
-                _NutriBadge(
-                  '${recipe.nutrition!.proteinG.round()}g',
-                  l10n.mealNutriProt,
-                  AppColors.primary(context),
-                ),
-                const SizedBox(width: 4),
-                _NutriBadge(
-                  '${recipe.nutrition!.carbsG.round()}g',
-                  l10n.mealNutriCarbs,
-                  AppColors.warning(context),
-                ),
-                const SizedBox(width: 4),
-                _NutriBadge(
-                  '${recipe.nutrition!.fatG.round()}g',
-                  l10n.mealNutriFat,
-                  AppColors.accent(context),
-                ),
-                if (recipe.nutrition!.fiberG >= 5) ...[
-                  const SizedBox(width: 4),
-                  _NutriBadge(
-                    '${recipe.nutrition!.fiberG.round()}g',
-                    l10n.mealNutriFiber,
-                    AppColors.success(context),
-                  ),
-                ],
-              ],
-            ),
-          ],
-          const SizedBox(height: 6),
-          // Action row: star rating + icon-only buttons
-          Row(
-            children: [
-              StarRatingRow(
-                currentRating: mealDay.effectiveRating,
-                onRate: onRating,
-              ),
-              const Spacer(),
-              SizedBox(
-                width: 32,
-                height: 32,
-                child: IconButton(
-                  onPressed: onToggleExpand,
-                  icon: Icon(
-                    isExpanded ? Icons.expand_less : Icons.restaurant_menu,
-                    size: 18,
-                  ),
-                  padding: EdgeInsets.zero,
-                  color: isExpanded
-                      ? AppColors.textSecondary(context)
-                      : AppColors.primary(context),
-                  tooltip: isExpanded ? l10n.close : l10n.mealIngredients,
-                ),
-              ),
-              SizedBox(
-                width: 32,
-                height: 32,
-                child: IconButton(
-                  onPressed: onViewPrepGuide,
-                  icon: const Icon(Icons.menu_book_outlined, size: 18),
-                  padding: EdgeInsets.zero,
-                  color: AppColors.primary(context),
-                  tooltip: l10n.mealViewPrepGuide,
-                ),
-              ),
-              SizedBox(
-                width: 32,
-                height: 32,
-                child: IconButton(
-                  onPressed: onSwap,
-                  icon: const Icon(Icons.swap_horiz, size: 18),
-                  padding: EdgeInsets.zero,
-                  color: AppColors.textSecondary(context),
-                  tooltip: l10n.mealSwap,
-                ),
-              ),
-            ],
-          ),
-          if (isExpanded) ...[
-            Divider(height: 1, color: AppColors.line(context)),
-            const SizedBox(height: 8),
-            // Show ingredients for ALL courses in this meal slot
-            ...allCourses.expand((course) {
-              final courseRecipe = service.recipeMap[course.recipeId];
-              if (courseRecipe == null) return <Widget>[];
-              final courseLabel = allCourses.length > 1
-                  ? _courseLabel(course.courseType, l10n)
-                  : null;
-              return [
-                if (courseLabel != null) ...[
-                  const SizedBox(height: 2),
-                  CalmEyebrow(courseLabel),
-                ] else ...[
-                  CalmEyebrow(l10n.mealIngredients.toUpperCase()),
-                ],
-                ...courseRecipe.ingredients.map((ri) {
-                  final effectiveId =
-                      course.substitutions[ri.ingredientId] ?? ri.ingredientId;
-                  final ing = iMap[effectiveId] ?? iMap[ri.ingredientId];
-                  if (ing == null) return const SizedBox();
-                  final isSubstituted = course.substitutions.containsKey(
-                    ri.ingredientId,
-                  );
-                  final dayGuests = plan.extraGuests[course.dayIndex] ?? 0;
-                  final scale =
-                      (plan.nPessoas + dayGuests) / courseRecipe.servings;
-                  final qty = ri.quantity * scale;
-                  final cost = qty * ing.avgPricePerUnit;
-                  // Tap → add to shopping list; tap icon → substitute ingredient.
-                  return CalmListTile(
-                    leadingIcon: isSubstituted ? Icons.swap_horiz : Icons.grain,
-                    leadingColor: isSubstituted
-                        ? AppColors.primary(context)
-                        : AppColors.textMuted(context),
-                    title: ing.name,
-                    subtitle:
-                        '${_fmt(qty)} ${ing.unit}${isSubstituted ? ' · substituted' : ''}',
-                    trailing: '${cost.toStringAsFixed(2)}${currencySymbol()}',
-                    onTap: () =>
-                        onSubstituteCourseIngredient(course, ri.ingredientId),
-                  );
-                }),
-              ];
-            }),
-            if (aiContent != null) ...[
-              const SizedBox(height: 8),
-              CalmEyebrow(l10n.mealPreparation.toUpperCase()),
-              const SizedBox(height: 4),
-              CalmCard(
-                padding: const EdgeInsets.all(12),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    ...aiContent!.steps.asMap().entries.map(
-                      (e) => Padding(
-                        padding: const EdgeInsets.only(bottom: 4),
-                        child: Text(
-                          '${e.key + 1}. ${e.value}',
-                          style: const TextStyle(fontSize: 13),
-                        ),
-                      ),
-                    ),
-                    if (aiContent!.tip.isNotEmpty) ...[
-                      Divider(height: 1, color: AppColors.line(context)),
-                      CalmListTile(
-                        leadingIcon: Icons.lightbulb_outline,
-                        leadingColor: AppColors.warning(context),
-                        title: aiContent!.tip,
-                      ),
-                    ],
-                    if (aiContent!.variation.isNotEmpty) ...[
-                      Divider(height: 1, color: AppColors.line(context)),
-                      CalmListTile(
-                        leadingIcon: Icons.shuffle,
-                        leadingColor: AppColors.primary(context),
-                        title: l10n.mealVariation,
-                        subtitle: aiContent!.variation,
-                      ),
-                    ],
-                    if (aiContent!.pairingSuggestion.isNotEmpty) ...[
-                      Divider(height: 1, color: AppColors.line(context)),
-                      CalmListTile(
-                        leadingIcon: Icons.restaurant,
-                        leadingColor: AppColors.success(context),
-                        title: l10n.mealPairing,
-                        subtitle: aiContent!.pairingSuggestion,
-                      ),
-                    ],
-                    if (aiContent!.storageInfo.isNotEmpty) ...[
-                      Divider(height: 1, color: AppColors.line(context)),
-                      CalmListTile(
-                        leadingIcon: Icons.kitchen,
-                        leadingColor: AppColors.textSecondary(context),
-                        title: l10n.mealStorage,
-                        subtitle: aiContent!.storageInfo,
-                      ),
-                    ],
-                    if (mealDay.isLeftover &&
-                        aiContent!.leftoverIdea.isNotEmpty) ...[
-                      Divider(height: 1, color: AppColors.line(context)),
-                      CalmListTile(
-                        leadingIcon: Icons.auto_fix_high,
-                        leadingColor: AppColors.warn(context),
-                        title: l10n.mealLeftoverIdea,
-                        subtitle: aiContent!.leftoverIdea,
-                      ),
-                    ],
-                  ],
-                ),
-              ),
-            ],
-          ],
-        ],
-      ),
-    );
-  }
-
-  String _courseLabel(CourseType ct, S l10n) {
-    switch (ct) {
-      case CourseType.soupOrStarter:
-        return l10n.mealCourseSoupStarter.toUpperCase();
-      case CourseType.mainCourse:
-        return l10n.mealCourseMain.toUpperCase();
-      case CourseType.dessert:
-        return l10n.mealCourseDessert.toUpperCase();
-    }
-  }
-
-  /// Builds course rows as CalmListTile for the compact header display.
-  List<Widget> _buildCourseListTiles(
-    BuildContext context,
-    S l10n,
-    Recipe mainRecipe,
-    Map<String, Ingredient> iMap,
-  ) {
-    final widgets = <Widget>[];
-    final hasCourses = allCourses.length > 1;
-
-    if (!hasCourses) {
-      // Single course: show as CalmListTile
-      widgets.add(
-        CalmListTile(
-          leadingIcon: Icons.restaurant,
-          leadingColor: AppColors.primary(context),
-          title: mainRecipe.name,
-          onTap: onSwap,
-        ),
-      );
-      return widgets;
-    }
-
-    // Multi-course: soup/starter, main, dessert each as CalmListTile
-    for (final course in allCourses) {
-      if (course.courseType == CourseType.soupOrStarter) {
-        final soupRecipe = service.recipeMap[course.recipeId];
-        if (soupRecipe != null) {
-          widgets.add(
-            CalmListTile(
-              leadingIcon: Icons.soup_kitchen,
-              leadingColor: AppColors.warn(context),
-              title: soupRecipe.name,
-              onTap: () => onSwapCourse(course),
-            ),
-          );
-        }
-      }
-    }
-
-    // Main course
-    widgets.add(
-      CalmListTile(
-        leadingIcon: Icons.restaurant,
-        leadingColor: AppColors.primary(context),
-        title: mainRecipe.name,
-        onTap: onSwap,
-      ),
-    );
-
-    for (final course in allCourses) {
-      if (course.courseType == CourseType.dessert) {
-        final dessertRecipe = service.recipeMap[course.recipeId];
-        if (dessertRecipe != null) {
-          widgets.add(
-            CalmListTile(
-              leadingIcon: Icons.icecream,
-              leadingColor: AppColors.accent(context),
-              title: dessertRecipe.name,
-              onTap: () => onSwapCourse(course),
-            ),
-          );
-        }
-      }
-    }
-
-    return widgets;
-  }
-
-  String _fmt(double v) {
-    if (v == v.roundToDouble()) return v.round().toString();
-    return v.toStringAsFixed(1);
-  }
-}
-
-// -- Swap Bottom Sheet --------------------------------------------------------
-
-class _SwapSheet extends StatefulWidget {
-  final String currentRecipeId;
-  final int nPessoas;
-  final Map<String, Ingredient> ingredientMap;
-  final MealPlannerService service;
-  final MealSettings mealSettings;
-  final MealType currentMealType;
-  final CourseType? courseType;
-  final void Function(String) onSelect;
-
-  const _SwapSheet({
-    required this.currentRecipeId,
-    required this.nPessoas,
-    required this.ingredientMap,
-    required this.service,
-    required this.mealSettings,
-    required this.currentMealType,
-    this.courseType,
-    required this.onSelect,
-  });
-
-  @override
-  State<_SwapSheet> createState() => _SwapSheetState();
-}
-
-class _SwapSheetState extends State<_SwapSheet> {
-  bool _showAllMealTypes = false;
-  late List<Recipe> _alternatives;
-
-  @override
-  void initState() {
-    super.initState();
-    _alternatives = widget.service.alternativesFor(
-      widget.currentRecipeId,
-      widget.nPessoas,
-      ms: widget.mealSettings,
-      courseType: widget.courseType,
-    );
-  }
-
-  void _refreshAlternatives() {
-    setState(() {
-      _alternatives = widget.service.alternativesFor(
-        widget.currentRecipeId,
-        widget.nPessoas,
-        ms: widget.mealSettings,
-        crossType: _showAllMealTypes,
-        courseType: widget.courseType,
-      );
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = S.of(context);
-    final currentRecipe = widget.service.recipeMap[widget.currentRecipeId];
-    final currentCost = currentRecipe != null
-        ? widget.service.recipeCost(
-            currentRecipe,
-            widget.nPessoas,
-            widget.ingredientMap,
-          )
-        : 0.0;
-
-    return SafeArea(
-      child: Padding(
-        padding: const EdgeInsets.all(20),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              l10n.mealAlternatives,
-              style: const TextStyle(fontSize: 17, fontWeight: FontWeight.w700),
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                FilterChip(
-                  label: Text(l10n.mealSwapSameType),
-                  selected: !_showAllMealTypes,
-                  onSelected: (_) {
-                    _showAllMealTypes = false;
-                    _refreshAlternatives();
-                  },
-                ),
-                const SizedBox(width: 8),
-                FilterChip(
-                  label: Text(l10n.mealSwapAllTypes),
-                  selected: _showAllMealTypes,
-                  onSelected: (_) {
-                    _showAllMealTypes = true;
-                    _refreshAlternatives();
-                  },
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            CalmCard(
-              padding: const EdgeInsets.all(0),
-              child: Column(
-                children: [
-                  ..._alternatives.take(6).map((r) {
-                    final cost = widget.service.recipeCost(
-                      r,
-                      widget.nPessoas,
-                      widget.ingredientMap,
-                    );
-                    final delta = cost - currentCost;
-                    final deltaStr = delta >= 0
-                        ? '+${delta.toStringAsFixed(2)}${currencySymbol()}'
-                        : '${delta.toStringAsFixed(2)}${currencySymbol()}';
-                    final deltaColor = delta > 0
-                        ? AppColors.bad(context)
-                        : AppColors.ok(context);
-                    // Show meal type badge for cross-type results via subtitle
-                    final isCrossType = !r.suitableMealTypes.contains(
-                      widget.currentMealType.name,
-                    );
-                    final subtitleParts = [
-                      l10n.mealTotalCost(cost.toStringAsFixed(2)),
-                      if (_showAllMealTypes && isCrossType)
-                        r.suitableMealTypes.first,
-                    ];
-                    return CalmListTile(
-                      leadingIcon: Icons.restaurant_menu,
-                      leadingColor: deltaColor,
-                      title: r.name,
-                      subtitle: subtitleParts.join(' · '),
-                      trailing: deltaStr,
-                      onTap: () {
-                        Navigator.pop(context);
-                        widget.onSelect(r.id);
-                      },
-                    );
-                  }),
-                ],
-              ),
-            ),
-            const SizedBox(height: 8),
-            SizedBox(
-              width: double.infinity,
-              child: TextButton(
-                onPressed: () => Navigator.pop(context),
-                child: Text(l10n.cancel),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// -- Consolidated List Sheet --------------------------------------------------
-
-class _ConsolidatedSheet extends StatelessWidget {
-  final Map<String, double> totals;
-  final Map<String, Ingredient> ingredientMap;
-  final void Function(ShoppingItem) onAddToShoppingList;
-
-  const _ConsolidatedSheet({
-    required this.totals,
-    required this.ingredientMap,
-    required this.onAddToShoppingList,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = S.of(context);
-    final grouped = <IngredientCategory, List<MapEntry<String, double>>>{};
-    for (final entry in totals.entries) {
-      final ing = ingredientMap[entry.key];
-      if (ing == null) continue;
-      grouped.putIfAbsent(ing.category, () => []).add(entry);
-    }
-
-    final categories = [
-      IngredientCategory.proteina,
-      IngredientCategory.vegetal,
-      IngredientCategory.carbo,
-      IngredientCategory.gordura,
-      IngredientCategory.condimento,
-    ];
-
-    return DraggableScrollableSheet(
-      initialChildSize: 0.7,
-      maxChildSize: 0.95,
-      expand: false,
-      builder: (_, controller) => Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 12),
-            child: ClipRRect(
-              // Justified: drag-handle indicator, no Calm equivalent
-              borderRadius: BorderRadius.circular(2),
-              child: SizedBox(
-                width: 40,
-                height: 4,
-                child: ColoredBox(color: AppColors.borderMuted(context)),
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 20),
-            child: Align(
-              alignment: Alignment.centerLeft,
-              child: Text(
-                l10n.mealConsolidatedTitle,
-                style: const TextStyle(
-                  fontSize: 17,
-                  fontWeight: FontWeight.w700,
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(height: 8),
-          Expanded(
-            child: ListView(
-              controller: controller,
-              padding: const EdgeInsets.fromLTRB(20, 0, 20, 20),
-              children: categories.expand((cat) {
-                final items = grouped[cat];
-                if (items == null || items.isEmpty) return <Widget>[];
-                return [
-                  const SizedBox(height: 16),
-                  CalmEyebrow(_categoryLabel(cat, l10n).toUpperCase()),
-                  const SizedBox(height: 4),
-                  CalmCard(
-                    padding: const EdgeInsets.all(0),
-                    child: Column(
-                      children: [
-                        ...items.map((entry) {
-                          final ing = ingredientMap[entry.key]!;
-                          final cost = entry.value * ing.avgPricePerUnit;
-                          return CalmListTile(
-                            leadingIcon: _ingredientCategoryIcon(cat),
-                            leadingColor: _ingredientCategoryColor(
-                              cat,
-                              context,
-                            ),
-                            title: ing.name,
-                            subtitle: '${_fmt(entry.value)} ${ing.unit}',
-                            trailing:
-                                '${cost.toStringAsFixed(2)}${currencySymbol()}',
-                            onTap: () => onAddToShoppingList(
-                              ShoppingItem(
-                                productName: ing.name,
-                                store: '',
-                                price: cost,
-                                unitPrice:
-                                    '${ing.avgPricePerUnit.toStringAsFixed(2)}${currencySymbol()}/${ing.unit}',
-                              ),
-                            ),
-                          );
-                        }),
-                      ],
-                    ),
-                  ),
-                ];
-              }).toList(),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  String _categoryLabel(IngredientCategory cat, S l10n) {
-    switch (cat) {
-      case IngredientCategory.proteina:
-        return l10n.mealCatProteins;
-      case IngredientCategory.vegetal:
-        return l10n.mealCatVegetables;
-      case IngredientCategory.carbo:
-        return l10n.mealCatCarbs;
-      case IngredientCategory.gordura:
-        return l10n.mealCatFats;
-      case IngredientCategory.condimento:
-        return l10n.mealCatCondiments;
-    }
-  }
-
-  IconData _ingredientCategoryIcon(IngredientCategory cat) {
-    switch (cat) {
-      case IngredientCategory.proteina:
-        return Icons.egg_outlined;
-      case IngredientCategory.vegetal:
-        return Icons.eco_outlined;
-      case IngredientCategory.carbo:
-        return Icons.grain;
-      case IngredientCategory.gordura:
-        return Icons.water_drop_outlined;
-      case IngredientCategory.condimento:
-        return Icons.spa_outlined;
-    }
-  }
-
-  Color _ingredientCategoryColor(IngredientCategory cat, BuildContext context) {
-    switch (cat) {
-      case IngredientCategory.proteina:
-        return AppColors.error(context);
-      case IngredientCategory.vegetal:
-        return AppColors.ok(context);
-      case IngredientCategory.carbo:
-        return AppColors.warning(context);
-      case IngredientCategory.gordura:
-        return AppColors.accent(context);
-      case IngredientCategory.condimento:
-        return AppColors.primary(context);
-    }
-  }
-
-  String _fmt(double v) {
-    if (v == v.roundToDouble()) return v.round().toString();
-    return v.toStringAsFixed(1);
-  }
-}
-
-// -- Helpers ------------------------------------------------------------------
-
-class _Stars extends StatelessWidget {
-  final int complexity;
-  const _Stars(this.complexity);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: List.generate(
-        5,
-        (i) => Icon(
-          i < complexity ? Icons.star : Icons.star_border,
-          size: 13,
-          color: AppColors.warning(context),
-        ),
-      ),
-    );
-  }
-}
-
-class _NutriBadge extends StatelessWidget {
-  final String value;
-  final String label;
-  final Color color;
-  const _NutriBadge(this.value, this.label, this.color);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      // Justified: micro nutrition badge — no Calm equivalent for inline tinted label
-      padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
-      decoration: BoxDecoration(
-        // Justified: micro nutrition badge colour tint
-        color: color.withValues(alpha: 0.08),
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Text(
-        '$value $label',
-        style: TextStyle(
-          fontSize: 10,
-          fontWeight: FontWeight.w600,
-          color: color,
-        ),
-      ),
-    );
-  }
-}
-
-class _WeeklySummaryCard extends StatelessWidget {
-  final WeeklyNutritionSummary summary;
-  const _WeeklySummaryCard({required this.summary});
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = S.of(context);
-    final scoreColor = summary.overallScore >= 7
-        ? AppColors.success(context)
-        : summary.overallScore >= 4
-        ? AppColors.warning(context)
-        : AppColors.error(context);
-
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16, 4, 16, 4),
-      child: CalmCard(
-        padding: const EdgeInsets.all(12),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Icon(
-                  Icons.analytics_outlined,
-                  size: 18,
-                  color: AppColors.ink70(context),
-                ),
-                const SizedBox(width: 8),
-                Text(
-                  l10n.mealWeeklySummary,
-                  style: const TextStyle(
-                    fontSize: 13,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-                const Spacer(),
-                CalmPill(
-                  label: '${summary.overallScore}/10',
-                  color: scoreColor,
-                ),
-              ],
-            ),
-            if (summary.highlights.isNotEmpty) ...[
-              const SizedBox(height: 8),
-              ...summary.highlights.map(
-                (h) => Padding(
-                  padding: const EdgeInsets.only(bottom: 3),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Icon(
-                        Icons.check_circle_outline,
-                        size: 14,
-                        color: AppColors.ok(context),
-                      ),
-                      const SizedBox(width: 6),
-                      Expanded(
-                        child: Text(h, style: const TextStyle(fontSize: 12)),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-            if (summary.concerns.isNotEmpty) ...[
-              const SizedBox(height: 4),
-              ...summary.concerns.map(
-                (c) => Padding(
-                  padding: const EdgeInsets.only(bottom: 3),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Icon(
-                        Icons.warning_amber_outlined,
-                        size: 14,
-                        color: AppColors.warn(context),
-                      ),
-                      const SizedBox(width: 6),
-                      Expanded(
-                        child: Text(c, style: const TextStyle(fontSize: 12)),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-          ],
-        ),
-      ),
-    );
   }
 }
